Purges bus recipients more precisely on `recurly.configure`
Introduction

Re-configuration of a recurly instance introduces a potential to abandon
hostedFields instances, say if the new configuration blob assigns a new
DOM target for one of the hosted fields. The routine to handle this is
contained within `recurly.parent`.

For a long time, it was sufficient to discard the recurly.bus when this
happened. The bus was safely considered to be dirty, and there were no
cases where a bus recipient may be lost in the discard procedure.

The Problem

However, with the introduction of Frame, and its newer dependents, we
came to add recipients to the bus that require the bus to persist across
the lifetime of their recurly instance.

Consider the following:

```js
const recurly = new Recurly();
const frame = recurly.Frame();
frame.attach(document.querySelector('#my-target'));
```

At this point, `frame` has added itself to `recurly.bus`, so that
messages that hit the parent window will dispatch to the `frame`
instance. Then let's say, while the frame is open, you do the following:

```js
recurly.configure({ fields: { card: { selector: '#new-selector' } } });
```

At this point, `recurly.parent` will destroy its `HostedFields` instance,
and create a new one targetting the new selector. It will then discard
`recurly.bus` and create a new one. This is a problem for our `frame`
instance, as it is listening to a bus that has been destroyed!

The Solution

This is a very sensitive aspect of a Recurly instance. The `recurly.bus`
is crucial for all messaging between Recurly.js classes, and
`recurly.parent` must be able to intelligently discard and re-create
HostedFields as configuration changes, and do so in a clean manner.

I determined that the bus was only being discarded in order to remove
any now-inactive hosted fields. Thus, it was just a matter of
instructing HostedField and HostedFields instances to remove themselves
from their bus when they are destroyed. This is a more clear and more
clean approach, and signals the true intent of the routine. No more
guessing! This had the added benefit of being able to remove the
long-maligned `Hostedfields.reset` method.

With Elements, this manner of shared state management is no longer a
concern, and situations like this are a perfect example of why that
architecture is an overall improvement.