Fixes Frame close race condition
When a Frame is created as a popover window, a window close listener
is created to observe when the window has been closed. This runs on a 1s
interval.

If a Frame subsequently finishes its task and sends its completion event,
that close listener was not properly disposed of. The resulting
behavior was a race condition where the close listener may announce
the window closure after the Frame had completed its task.

This ensures the close listener is cleared, and ends the race behavior
when the Frame is done.