Restructures types for DT
We restructured the types to comply with structural expectations
in DefinitelyTyped
(https://github.com/DefinitelyTyped/DefinitelyTyped/pull/43900)

This is a follow up to https://github.com/recurly/recurly-js/pull/591
which exports types.

All types are now defined in lib. The index.d.ts declaration file is
only responsible for declaring recurly as a global variable and
exporting types.