Patches dtslint config to match DefinitelyTyped
1. Adds linting for /types.

Previously, we only ran tests on test/types, but DefinitelyTyped is also
linting the types themselves, not just the tests, resulting in potential
undiscovered linting errors emerging when opening our PR to
DefinitelyTyped
(https://github.com/DefinitelyTyped/DefinitelyTyped/pull/45675/checks?check_run_id=801301793)

To fix this, we're now linting our types with dtslint. We had to move
types/ to DefinitelyTyped/types/recurly__recurly-js to circumvent this
error
https://github.com/microsoft/dtslint/blob/master/src/index.ts#L214-L227

This was also added as a CI step.

2. Fixes test/tslint.json to match dtslint

We had some discrepancies in our tsconfig.compilerOptions that caused
typescript to behave differently. We updated test/tslint.json to match
the one on DefinitelyTyped.

3. Fixes linting and testing errors from updated config

The errors being addressed are the same errors on DefinitelyTyped
causing CI to fail.