feat(apple pay): populate billing/shippingContact with form address fields
Apple Pay on the Web supports supplying a `billingContact` and
`shippingContact` when creating the session. If provided, it would
populate over the customer's default billing contact on the payment
sheet.

If the `form` option is supplied, use the address fields that are
populated to populate the `billingContact` and the shipping fields
that are populated to populate the `shippingContact`.

This also cleans up the interaction between the `form` and the token.
Currently, if any of the fields on the `form` are present, we prefer
that address instead of what is presented and modified on the payment
card. This changes that functionality to always prefer the
`billingContact` from the payment card when tokenizing the card as
that is what the customer sees when authorizing the payment.