feat(apple pay): Move payment request properties to `options.paymentRequest`
Cleans up the interface for calling into `recurly.ApplePay` so that the
properties that drift down to the resulting payment request are not
intermixed with the SDK specific parameters.

As a result, deprecates `options.requiredShippingContactFields` in favor
of reading it off of the `paymentRequest.requiredShippingContactFields`.

Adds `options.recurring` that, when provided in combination with
`options.total`, will create a total that will include a monthly
recurring charge on the payment sheet.

`options.paymentRequest` will be treated as the source of truth, except
for when `options.pricing` is provided. This maintains the stance that
Pricing controls the `total` and `lineItems`.