fix(apple pay): Do not teardown Braintree on cancel
`teardown` completely removes the Braintree instance which will cause
issues if the client cancels and tries to initialize a payment. This is
not needed as beginning a session does not reinitialize the BT client,
so we can have successive attempts on the same client.

Fixes the `performValidation cannot be called after teardown` error on
subsequent tries.