fix(apple pay): callbacks with promises resolve correctly
When a callback returns a promise, R.js was using `finally` to update
the payment sheet, however `finally` does not accept any arguments so no
updates were actually propagated through to the payment sheet. This
fixes that by using a `catch.then` to ensure that the promise is handled
value is propagated correctly.