fix(apple pay): simplify newRecurringPaymentRequest handling
There are a variety of methods to initialize ApplePay with various
levels of integration with the native ApplePay request. On the simplest
level of configuration (`total`, `recurring` options), the merchant
might want to register the `onPaymentMethodSelected` callback to update
the total. Since the library creates the `recurringPaymentRequest` for
them automatically, we need to keep that total up to date if it changes.

We also manage the managementURL, so that property can be omitted when
using the callback and added in automatically by the library.