chore: convert Google Pay element to a class from function
To better align the Google Pay element with the Apple Pay element, this
refactors it to be a concrete class instead of a function. This will
help in the future when we need to specialize its functionality for a
different implementation of collecting the payment information. Adding
Braintree support is an example of that specialization.